Move getUserTag/getMention, refactor isPremium,other extracte methods
