+++ src/commands/game_commands/forceskip.ts
@@ -4,7 +4,6 @@ import {
     areUserAndBotInSameVoiceChannel,
     getDebugLogHeader,
     sendInfoMessage,
-    getMention,
 } from "../../helpers/discord_utils";
 import { getGuildPreference } from "../../helpers/game_utils";
 import { IPCLogger } from "../../logger";
@@ -15,6 +14,7 @@ import Session from "../../structures/session";
 import type CommandArgs from "../../interfaces/command_args";
 import type HelpDocumentation from "../../interfaces/help";
 import LocalizationManager from "../../helpers/localization_manager";
+import { getMention } from "../../helpers/utils";
 
 const logger = new IPCLogger("forceskip");
