+++ src/seed/seed_db.ts
@@ -8,7 +8,6 @@ import { IPCLogger } from "../logger";
 import { downloadAndConvertSongs } from "../scripts/download-new-songs";
 import type { DatabaseContext } from "../database_context";
 import { getNewConnection } from "../database_context";
-import { generateKmqDataTables, loadStoredProcedures } from "./bootstrap";
 import _ from "lodash";
 import { parseJsonFile } from "../helpers/utils";
 import { EnvType } from "../enums/env_type";
@@ -52,6 +51,35 @@ async function getOverrideQueries(db: DatabaseContext): Promise<Array<string>> {
     return (await db.kmq("kpop_videos_sql_overrides").select(["query"])).map(
         (x) => x.query
     );
+}
+
+// eslint-disable-next-line import/prefer-default-export
+/**
+ * Re-creates the KMQ data tables
+ * @param db - The database context
+ */
+export async function generateKmqDataTables(
+    db: DatabaseContext
+): Promise<void> {
+    logger.info("Re-creating KMQ data tables view...");
+    await db.kmq.raw(
+        `CALL CreateKmqDataTables(${process.env.PREMIUM_AUDIO_SONGS_PER_ARTIST});`
+    );
+}
+
+/**
+ * Reloads all existing stored procedures
+ */
+export function loadStoredProcedures(): void {
+    const storedProcedureDefinitions = fs
+        .readdirSync(path.join(__dirname, "../../sql/procedures"))
+        .map((x) => path.join(__dirname, "../../sql/procedures", x));
+
+    for (const storedProcedureDefinition of storedProcedureDefinitions) {
+        execSync(
+            `mysql -u ${process.env.DB_USER} -p${process.env.DB_PASS} -h ${process.env.DB_HOST} --port ${process.env.DB_PORT} kmq < ${storedProcedureDefinition}`
+        );
+    }
 }
 
 const downloadDb = async (): Promise<void> => {
